+++ src/main/java/presentation/ArtistsController.java
@@ -57,7 +57,7 @@ public class ArtistsController implements Initializable {
     @FXML
     void deleteArtist(MouseEvent event) {
         int index = artists.getSelectionModel().getFocusedIndex();
-        long userID = artistsObservableList.get(index).getId();
+        int userID = artistsObservableList.get(index).getId();
 
         if (artistsObservableList.get(index).getId() == 6){
             return;
